# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Knotree Accounts API
  version: 0.1.0
servers:
  - url: https://accounts.knotree.com
paths:
...
  /oauth/authorize:
    get:
      summary: OAuth 2.1 authorization endpoint
      parameters:
        - in: query
          name: client_id
          required: true
          schema: { type: string }
        - in: query
          name: redirect_uri
          required: true
          schema: { type: string }
        - in: query
          name: response_type
          required: true
          schema: { type: string, enum: [code] }
        - in: query
          name: code_challenge
          required: true
          schema: { type: string }
        - in: query
          name: code_challenge_method
          required: true
          schema: { type: string, enum: [S256] }
        - in: query
          name: state
          required: true
          schema: { type: string }
      responses:
        "302": { description: Redirect to the client or to sign-in }
        "400": { description: Invalid client or redirect URI }
  /oauth/token:
    post:
      summary: Token endpoint
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                grant_type: { type: string }
                code: { type: string }
                redirect_uri: { type: string }
                client_id: { type: string }
                code_verifier: { type: string }
                refresh_token: { type: string }
      responses:
        "200": { description: Tokens issued }
        "400": { description: Invalid grant }
...
components:
  securitySchemes:
    browserSession:
      type: apiKey
      in: cookie
      name: knotree_session
    bearer:
      type: http
      scheme: bearer